backend/requirements.txt
flask
flask-cors
flask-jwt-extended
psycopg2-binary
python-dotenv
bcrypt
google-generativeai
tensorflow-cpu
numpy
pandas
scikit-learn
gunicorn
requests
